Fix App Crash on ThreadCard click event.
App crashes when ThreadCard is clicked.
Caused primarily due to serialization of
MappedByteBuffer, which is not supported.
Fix:
Remove the need for pushing intent for
"detector" to the ThreadActivity, since
individual messages have already been
classified when they are added to the list.
Thus also removing the excess unnecessary code.