Update LandingPageViewModel
1. Add messageIndex: Dictionary containing
all the messages indexed by their ID.
2. loadThreads() will call loadMessageIndex()
to first gather all the messages either from
the data store or if none stored, then
query the content resolver.
3. refreshSmsMessages(): As a temporary solution,
if will check if the previous list size is
different from the new size, if the condition
is true, then it will load the threads.
4. runPredictor(): If no API_URL is set,
then exit the method. Otherwise, for each message
call the API and update the data store.

All the above methods will have to refactored
as their overhead of the re-running the same
thing over and again.