Update the column type for spam.
Instead of integer, use text.
It was creating problems during
loading, as null type was defaulting
to 0 which is false, which means
new messages stored and not identified
were stored with NULL spam type,
however during loading from the dataStore
the NULL spam type, was converted to
0 due to integer constraint, thus
the making the NULL type 'false',
thus making the message "not spam".